\name{bestSPLS-package}
\alias{bestSPLS-package}
\alias{bestSPLS}
\docType{package}
\title{
\packageTitle{bestSPLS}
}
\description{
\packageDescription{bestSPLS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bestSPLS}
\packageIndices{bestSPLS}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bestSPLS}

Maintainer: \packageMaintainer{bestSPLS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
